Moved navigation menu over from personal project to SofEng share

+++ db/migrate/20141027210621_devise_create_users.rb
@@ -8,10 +8,10 @@ class DeviseCreateUsers < ActiveRecord::Migration
       t.integer :u_type
       t.text :gender
       t.datetime :bday
-      t.double :weight
-      t.double :height
-      t.double :ex_score
-      t.double :diet_score
+      t.integer :weight
+      t.integer :height
+      t.integer :ex_score
+      t.integer :diet_score
       t.boolean :diet_fitness_flag
 
       ## Database authenticatable
